+++ CHANGELOG.md
@@ -17,6 +17,7 @@ _2023_
 
 - LBA1 Climbing Ladders
 - LBA1 3D Model Shading
+- LBA1 3D Model Spheres and Lines Colours
 - Extra bonus speed
 - Behavioiur Menu Icons CSS
 - Fixed Sprites out of place caused by wrong scene point coordinates